back-end/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-stubs==5.0.0
django-stubs-ext==5.0.0
django-types==0.19.1
djangorestframework==3.15.1
djangorestframework-types==0.8.0
iniconfig==2.0.0
mypy==1.10.0
mypy-extensions==1.0.0
packaging==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.0
python-dotenv==1.0.1
sqlparse==0.5.0
types-psycopg2==2.9.21.20240417
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
